support for subfolders in zip files
Former-commit-id: c01c933fe81e6bfa0ebe36d1e7f400f33211fbde [formerly 18ab8b82b86424546749bde869034ec56002f48a] [formerly ed04e890332802e70fe755e4dd77bcc149b4c94d [formerly fc3afedaeed8faf6e938dd9120a0140ae8e8835c]] [formerly 7975dd9fef6fb0566f6c7410e2e3309656e219ee [formerly 0b3815953bdf4f79fd231adeda9b05fcabb9e9eb] [formerly 12c24ea199062e35c5b00119390ebf7e21ee77b5 [formerly 510b37c5b07173184682133f5ea7c16b772fbe72]]]
Former-commit-id: f9b498a5fdb9122615f600f9c8e26a2cc9104562 [formerly d6ff0edbb5478ef4cb85a5ca078d51cc6c1cb54e] [formerly 46a3628049b4a571f725f0fd9314b9404f7d8a32 [formerly ef86c2c098f4a33b6a813eaac4e093b191986cc6]]
Former-commit-id: e7735b6da4ccbae866927de259324a965a5bfec2 [formerly 78ef9784461da9b4e157450c223c63bd5a2a48a1]
Former-commit-id: 82a670a37b69cc443d21f9c1b9b8dd15d5a7b206

+++ pycorrfit/readfiles/__init__.py
@@ -26,8 +26,10 @@
 import csv
 import numpy as np
 import os
+import sys
 import tempfile
 import yaml
+import warnings
 import zipfile
 
 # To add a filetype add it here and in the
@@ -110,8 +112,8 @@ def openZIP(dirname, filename):
     Trace = list()        # Corresponding traces
     ## First test, if we are opening a session file
     sessionwc = [".fcsfit-session.zip", ".pcfs"]
-    if ( (len(filename)>19 and filename[-19:] == sessionwc[0]) or
-         (len(filename)> 5 and filename[-5:] == sessionwc[1])     ):
+    if ( filename.endswith(sessionwc[0]) or
+         filename.endswith(sessionwc[1])     ):
         # Get the yaml parms dump:
         yamlfile = Arc.open("Parameters.yaml")
         # Parms: Fitting and drawing parameters of the correlation curve
@@ -198,22 +200,33 @@ def openZIP(dirname, filename):
         allfiles = Arc.namelist()
         # Extract data to temporary folder
         tempdir = tempfile.mkdtemp()
+        rmdirs = list()
         for afile in allfiles:
-            Arc.extract(afile, path=tempdir)
+            apath = Arc.extract(afile, path=tempdir)
+            if os.path.isdir(apath):
+                rmdirs.append(apath)
+                continue
             ReturnValue = openAny(tempdir, afile)
             if ReturnValue is not None:
-                cs = ReturnValue["Correlation"]
-                ts = ReturnValue["Trace"]
-                ls = ReturnValue["Type"]
-                fs = ReturnValue["Filename"]
-                for i in np.arange(len(cs)):
-                    Correlations.append(cs[i])
-                    Trace.append(ts[i])
-                    Curvelist.append(ls[i])
-                    Filelist.append(filename+"/"+fs[i])
+                Correlations += ReturnValue["Correlation"]
+                Trace += ReturnValue["Trace"]
+                Curvelist += ReturnValue["Type"]
+                fnames = ReturnValue["Filename"]
+                Filelist += [ filename+"/"+fs for fs in fnames ]
             # Delte file
-            os.remove(os.path.join(tempdir,afile))
-        os.removedirs(tempdir)
+            try:
+                os.remove(os.path.join(tempdir, afile))
+            except:
+                warnings.warn("{}".format(sys.exc_info()[1]))
+        for rmd in rmdirs:
+            try:
+                os.removedirs(rmd)
+            except:
+                    warnings.warn("{}".format(sys.exc_info()[1]))
+        try:
+            os.removedirs(tempdir)
+        except:
+                warnings.warn("{}".format(sys.exc_info()[1]))
     Arc.close()
     dictionary = dict()
     dictionary["Correlation"] = Correlations
